[^+] Prepared for release
  [+] [License]
  [+] [Readme]
  [+] [Install/Remove]

  [-] [Interface] Moved Test() to example
  [/] [Interface] Renamed to "TRgbStripe", not "TRgbStripeManager"
  [~^] [Example] Revised code and readability

Well, you should write documentation just before release.

At writing example you understand that Test() should belong not to class
but to example. And that "Manager" is design implementation pattern,
this word is not important for practical use.

So a better name for usage is like "RGB stripe" or just "Stripe".